+++ Podfile
@@ -8,6 +8,7 @@ target 'OneDayOneAnswer' do
 	# Pods for OneDayOneAnswer
 	# add the Firebase pod for Google Analytics
 	
+	pod 'Firebase/Crashlytics'
 	pod 'Firebase/Analytics'
 	pod 'Firebase/Auth'
 	pod 'GoogleSignIn'
